requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
azure-core==1.30.1
beautifulsoup4==4.12.3
blinker==1.7.0
botframework-connector==4.14.8
botframework-streaming==4.14.8
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.1
databases==0.9.0
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.1
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.0.2
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpickle==1.4.2
lxml==5.1.0
MarkupSafe==2.1.5
msal==1.28.0
msrest==0.7.1
multidict==6.0.5
ngrok==0.12.1
oauthlib==3.2.2
openai==1.25.2
packaging==23.2
pdfminer.six==20231228
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
PyJWT==2.8.0
pyngrok==7.1.2
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.1
python-pptx==0.6.23
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.26.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
tqdm==4.66.4
typing_extensions==4.9.0
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.29.0
Werkzeug==3.0.1
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.17.0
